+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Enable metrics collection with default interval of 30s (#322)
  - `ELASTIC_APM_SERVER_CERT` enables server certificate pinning (#325)
  - Add Docker container ID to metadata (#330)
+ - Added distributed trace context propagation to apmgrpc (#335)
 
 ## [v1.0.0](https://github.com/elastic/apm-agent-go/releases/tag/v1.0.0)